improvements on random e2e test failures
