+++ src/e2e-test/features/wrangler.bigquery.feature
@@ -19,25 +19,25 @@ Feature: Wrangler BigQuery Connection Tests
 
   @WRANGLER_BIGQUERY_CONNECTION_TEST
   Scenario: Should successfully test BIGQUERY connection
-    When Open Connections page for BigQuery connection test
+    When Open Connections Page
     Then Test BigQuery connection with name "bigquery_randomtest1234"
     Then Verify for successful BigQuery test connection and message
 
   @WRANGLER_BIGQUERY_CONNECTION_TEST
   Scenario: Should show appropriate message when BigQuery connection fails
-    When Open Connections page
+    When Open Connections Page
     Then Test BigQuery connection with "unknown_bigquery_connection_name", "unknown_project", "unknown_path"
     Then Verify BigQuery test connection failure and message
 
   @WRANGLER_BIGQUERY_CONNECTION_TEST
   Scenario: Should create BIGQUERY connection
-    When Open Connections page
+    When Open Connections Page
     Then Create BigQuery connection with name "bigquery_randomtest1234"
     Then Verify navigation to created BigQuery connection "bigquery_randomtest1234"
 
   @WRANGLER_BIGQUERY_CONNECTION_TEST
   Scenario: Should show proper error message when trying to create existing BigQuery connection
-    When Open Connections page
+    When Open Connections Page
     Then Create BigQuery connection with name "bigquery_randomtest1234"
     Then Check for the BigQuery connection already exists error for "bigquery_randomtest1234"
 
@@ -48,14 +48,14 @@ Feature: Wrangler BigQuery Connection Tests
 
   @WRANGLER_BIGQUERY_CONNECTION_TEST
   Scenario: Should be able navigate inside BIGQUERY connection & create workspace
-    When Open Connections page
+    When Open Connections Page
     Then Select BigQuery connection "bigquery_randomtest1234"
     Then Check BigQuery connection "bigquery_randomtest1234" details: Instance, Database, Table
     Then Verify URL navigation for BigQuery connection
 
   @WRANGLER_BIGQUERY_CONNECTION_TEST
   Scenario: Should delete an existing connection
-    When Open Connections page
+    When Open Connections Page
     Then Delete BigQuery connection "bigquery_randomtest1234"
     Then Confirm that the BigQuery connection "bigquery_randomtest1234" is deleted
