+++ src/e2e-test/features/wrangler.gcs.feature
@@ -19,25 +19,25 @@ Feature: Wrangler GCS Connection Tests
 
   @WRANGLER_GCS_CONNECTION_TEST
   Scenario: Should successfully test GCS connection
-    When Open Connections page for GCS connection test
+    When Open Connections Page
     Then Test GCS connection with name "gcs_randomtest1234"
     Then Verify for successful GCS test connection and message
 
   @WRANGLER_GCS_CONNECTION_TEST
   Scenario: Should show appropriate message when GCS connection fails
-    When Open Connections page
+    When Open Connections Page
     Then Test GCS connection with "unknown_gcs_connection_name", "unknown_project", "unknown_path"
     Then Verify GCS test connection failure and message
 
   @WRANGLER_GCS_CONNECTION_TEST
   Scenario: Should create GCS connection
-    When Open Connections page
+    When Open Connections Page
     Then Create GCS connection with name "gcs_randomtest1234"
     Then Verify navigation to created GCS connection "gcs_randomtest1234"
 
   @WRANGLER_GCS_CONNECTION_TEST
   Scenario: Should show proper error message when trying to create existing GCS connection
-    When Open Connections page
+    When Open Connections Page
     Then Create GCS connection with name "gcs_randomtest1234"
     Then Check for the GCS connection already exists error for "gcs_randomtest1234"
 
@@ -48,14 +48,14 @@ Feature: Wrangler GCS Connection Tests
 
   @WRANGLER_GCS_CONNECTION_TEST
   Scenario: Should be able navigate inside GCS connection & create workspace
-    When Open Connections page
+    When Open Connections Page
     Then Select GCS connection "gcs_randomtest1234"
     Then Check GCS connection "gcs_randomtest1234" details: Folder, file
     Then Verify URL navigation for GCS connection
 
   @WRANGLER_GCS_CONNECTION_TEST
   Scenario: Should delete an existing connection
-    When Open Connections page
+    When Open Connections Page
     Then Delete GCS connection "gcs_randomtest1234"
     Then Confirm that the GCS connection "gcs_randomtest1234" is deleted
