+++ src/e2e-test/features/wrangler.spanner.feature
@@ -19,25 +19,25 @@ Feature: Wrangler Spanner Tests
 
   @WRANGLER_SPANNER_TEST
   Scenario: Should successfully test SPANNER connection
-    When Open Connections page
+    When Open Connections Page
     Then Test Spanner Connection with name "spanner_randomtest1234"
     Then Verify for successful test connection and message
 
   @WRANGLER_SPANNER_TEST
   Scenario: Should show appropriate message when test connection fails
-    When Open Connections page
+    When Open Connections Page
     Then Test Spanner Connection with "unknown_spanner_connection_name", "unknown_project", "unknown_path"
     Then Verify test connection failure and message
 
   @WRANGLER_SPANNER_TEST
   Scenario: Should create SPANNER connection
-    When Open Connections page
+    When Open Connections Page
     Then Create Spanner Connection with name "spanner_randomtest1234"
     Then Verify navigation to created connection "spanner_randomtest1234"
 
   @WRANGLER_SPANNER_TEST
   Scenario: Should show proper error message when trying to create existing connection
-    When Open Connections page
+    When Open Connections Page
     Then Create Spanner Connection with name "spanner_randomtest1234"
     Then Check for the connection already exists error for "spanner_randomtest1234"
 
@@ -48,14 +48,14 @@ Feature: Wrangler Spanner Tests
 
   @WRANGLER_SPANNER_TEST
   Scenario: Should be able navigate inside SPANNER connection & create workspace
-    When Open Connections page
+    When Open Connections Page
     Then Select Spanner connection "spanner_randomtest1234"
     Then Check Connection "spanner_randomtest1234" details: Instance, Database, Table
     Then Verify URL navigation
 
   @WRANGLER_SPANNER_TEST
   Scenario: Should delete an existing connection
-    When Open Connections page
+    When Open Connections Page
     Then Delete Connection "spanner_randomtest1234"
     Then Confirm that the connection "spanner_randomtest1234" is deleted
