+++ src/e2e-test/java/io/cdap/cdap/ui/utils/Constants.java
@@ -74,4 +74,11 @@ public class Constants {
 
   public static final String SINK_PATH_VAL = "/tmp/cdap-ui-integration-fixtures";
   public static final String SOURCE_PATH_VAL = "file:/tmp/cdap-ui-integration-fixtures/airports.csv";
+
+  public static final String SOURCE_PLUGINS_GROUP_LOCATOR_TEXT = "Source";
+  public static final String SINK_PLUGINS_GROUP_LOCATOR_TEXT = "Sink";
+  public static final String TRANSFORM_PLUGINS_GROUP_LOCATOR_TEXT = "Transform";
+  public static final String ANALYTICS_PLUGINS_GROUP_LOCATOR_TEXT = "Analytics";
+  public static final String CONDITIONS_AND_ACTIONS_PLUGINS_GROUP_LOCATOR_TEXT = "Conditions and Actions";
+  public static final String ERROR_HANDLERS_AND_ALERTS_PLUGINS_GROUP_LOCATOR_TEXT = "Error Handlers and Alerts";
 }
